Ai-Waste-Management-System-Final-project--1/aiml/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
python-multipart>=0.0.6
pyyaml>=6.0
Pillow>=10.0.0
